user result file restructure
